auth/jwt_auth.py
from datetime import datetime, timedelta
import os

# Use PyJWT instead of python-jose to avoid conflicts
try:
    import jwt as pyjwt
    JWTError = pyjwt.InvalidTokenError
except ImportError:
    raise ImportError(
        "PyJWT is not installed. Please run: pip install PyJWT"
    )

SECRET_KEY = os.environ.get('JWT_SECRET_KEY', 'your-secret-key-change-in-production')
ALGORITHM = 'HS256'
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24 * 7  # 7 days

def create_access_token(data: dict, expires_delta: timedelta = None):
    """Create JWT access token"""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({'exp': expire})
    encoded_jwt = pyjwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def verify_token(token: str):
    """Verify JWT token and return payload"""
    try:
        payload = pyjwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except (JWTError, pyjwt.ExpiredSignatureError, pyjwt.InvalidTokenError):
        return None

