requirements.txt
playwright>=1.40.0
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0

